frontend: router: Extract getDefaultRoutes and Router to own files
Also add a setDefaultRoutes function so that people can import
getDefaultRoutes without depending on the main big routes file.
