+++ frontend/src/plugin/registry.tsx
@@ -40,7 +40,7 @@ import { setSidebarItem, setSidebarItemFilter } from '../components/Sidebar/side
 import { getHeadlampAPIHeaders } from '../helpers/getHeadlampAPIHeaders';
 import { AppTheme } from '../lib/AppTheme';
 import { KubeObject } from '../lib/k8s/KubeObject';
-import { Route } from '../lib/router';
+import type { Route } from '../lib/router/Route';
 import {
   addDetailsViewHeaderActionsProcessor,
   AppBarAction,
